pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nodrive-gpm-package"
version = "1.0.0"
description = "Browser Profile Management with Anti-Detection"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["gpm", "browser", "automation", "anti-detection", "nodriver"]
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "nodriver>=0.28",
    "requests>=2.31.0",
    "psutil>=5.9.0",
    "pywin32>=306",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
optional = [
    "python-dotenv>=1.0.0",
    "PySocks>=1.7.1",
]

[project.urls]
Homepage = "https://github.com/yourusername/nodrive-gpm-package"
Issues = "https://github.com/yourusername/nodrive-gpm-package/issues"
